requirements.txt
certifi==2021.10.8
chardet==4.0.0
idna==2.10
python-dotenv==0.18.0
requests==2.27.1
typing==3.10.0.0
urllib3==1.26.19
